Adding controls component for user film actions
