+++ project/src/components/film-promo/film-promo.tsx
@@ -1,10 +1,10 @@
 import Film from '../../types/film';
+import FilmControls from '../film-controls/film-controls';
 import Header from '../header/header';
 import { useAppSelector } from '../../hooks';
 
 function FilmPromo(): JSX.Element {
   const film = useAppSelector((state) => state.promo) as Film;
-  const favouritesCount = useAppSelector((state) => state.favouriteFilmList.length);
   const isPromoLoaded = useAppSelector((state) => state.isLoadingCompleted.promo);
 
   if (!isPromoLoaded) {
@@ -39,21 +39,7 @@ function FilmPromo(): JSX.Element {
               <span className="film-card__year">{film.year}</span>
             </p>
 
-            <div className="film-card__buttons">
-              <button className="btn btn--play film-card__button" type="button">
-                <svg viewBox="0 0 19 19" width="19" height="19">
-                  <use xlinkHref="#play-s"></use>
-                </svg>
-                <span>Play</span>
-              </button>
-              <button className="btn btn--list film-card__button" type="button">
-                <svg viewBox="0 0 19 20" width="19" height="20">
-                  <use xlinkHref="#add"></use>
-                </svg>
-                <span>My list</span>
-                <span className="film-card__count">{favouritesCount}</span>
-              </button>
-            </div>
+            <FilmControls film={film}/>
           </div>
         </div>
       </div>
