+++ project/src/pages/movie-screen/movie-screen.tsx
@@ -1,9 +1,8 @@
-import { APIRoute, AppRoute, AuthStatus } from '../../const';
-import { Link, useLocation } from 'react-router-dom';
-import { useAppDispatch, useAppSelector } from '../../hooks';
+import { APIRoute, AppRoute } from '../../const';
 import { useEffect, useState } from 'react';
 
 import Film from '../../types/film';
+import FilmControls from '../../components/film-controls/film-controls';
 import { FilmState } from '../../types/interface';
 import Footer from '../../components/footer/footer';
 import Header from '../../components/header/header';
@@ -13,13 +12,13 @@ import Tabs from '../../components/tabs/tabs';
 import api from '../../services/api';
 import { filmFromApi } from '../../services/adapters/film';
 import { redirectToRoute } from '../../store/action';
+import { useAppDispatch } from '../../hooks';
+import { useLocation } from 'react-router-dom';
 
 function MovieScreen(): JSX.Element {
   const { filmID } = useLocation().state as FilmState;
   const [film, setFilmData] = useState({} as Film);
   const [isLoadingCompleted, setLoadingCompleted] = useState(false);
-  const { pathname } = useLocation();
-  const authStatus = useAppSelector((state) => state.authorizationStatus);
   const dispatch = useAppDispatch();
 
   useEffect(() => {
@@ -63,28 +62,7 @@ function MovieScreen(): JSX.Element {
                 <span className="film-card__year">{film.year}</span>
               </p>
 
-              <div className="film-card__buttons">
-                <button className="btn btn--play film-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list film-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <span>My list</span>
-                  <span className="film-card__count">9</span>
-                </button>
-                { authStatus === AuthStatus.Auth &&
-                <Link
-                  className="btn film-card__button"
-                  to={`${pathname}/review`}
-                  state={{ film: film }}
-                >Add review
-                </Link> }
-              </div>
+              <FilmControls film={film}/>
             </div>
           </div>
         </div>
@@ -92,7 +70,12 @@ function MovieScreen(): JSX.Element {
         <div className="film-card__wrap film-card__translate-top">
           <div className="film-card__info">
             <div className="film-card__poster film-card__poster--big">
-              <img src={film.posterSrc} alt={`${film.title} poster`} width="218" height="327" />
+              <img
+                src={film.posterSrc}
+                alt={`${film.title} poster`}
+                width="218"
+                height="327"
+              />
             </div>
 
             <Tabs film={film}/>
